refactor: use nested routes for home and explore page
