+++ apps/ui/src/views/My.vue
@@ -1,0 +1,17 @@
+<script setup lang="ts">
+const uiStore = useUiStore();
+</script>
+
+<template>
+  <div>
+    <div>
+      <div
+        class="ml-0 lg:ml-[240px] mr-0 xl:mr-[240px]"
+        :class="{ 'translate-x-[240px] lg:translate-x-0': uiStore.sidebarOpen }"
+      >
+        <router-view />
+      </div>
+      <div class="invisible xl:visible fixed w-[240px] border-l bottom-0 top-[72px] right-0" />
+    </div>
+  </div>
+</template>
