+++ apps/ui/src/views/My/Home.vue
@@ -110,54 +110,48 @@ watch(filter, (toFilter, fromFilter) => {
 
 <template>
   <div>
-    <div
-      class="ml-0 lg:ml-[240px] mr-0 xl:mr-[240px]"
-      :class="{ 'translate-x-[240px] lg:translate-x-0': uiStore.sidebarOpen }"
-    >
-      <div class="flex justify-between">
-        <div class="flex flex-row p-4 space-x-2">
-          <UiSelectDropdown
-            v-model="filter"
-            title="Status"
-            gap="12px"
-            placement="left"
-            :items="[
-              {
-                key: 'any',
-                label: 'Any'
-              },
-              {
-                key: 'pending',
-                label: 'Pending',
-                component: ProposalIconStatus,
-                componentProps: { ...selectIconBaseProps, state: 'pending' }
-              },
-              {
-                key: 'active',
-                label: 'Active',
-                component: ProposalIconStatus,
-                componentProps: { ...selectIconBaseProps, state: 'active' }
-              },
-              {
-                key: 'closed',
-                label: 'Closed',
-                component: ProposalIconStatus,
-                componentProps: { ...selectIconBaseProps, state: 'passed' }
-              }
-            ]"
-          />
-        </div>
+    <div class="flex justify-between">
+      <div class="flex flex-row p-4 space-x-2">
+        <UiSelectDropdown
+          v-model="filter"
+          title="Status"
+          gap="12px"
+          placement="left"
+          :items="[
+            {
+              key: 'any',
+              label: 'Any'
+            },
+            {
+              key: 'pending',
+              label: 'Pending',
+              component: ProposalIconStatus,
+              componentProps: { ...selectIconBaseProps, state: 'pending' }
+            },
+            {
+              key: 'active',
+              label: 'Active',
+              component: ProposalIconStatus,
+              componentProps: { ...selectIconBaseProps, state: 'active' }
+            },
+            {
+              key: 'closed',
+              label: 'Closed',
+              component: ProposalIconStatus,
+              componentProps: { ...selectIconBaseProps, state: 'passed' }
+            }
+          ]"
+        />
       </div>
-      <ProposalsList
-        title="Proposals"
-        limit="off"
-        :loading="!loaded"
-        :loading-more="loadingMore"
-        :proposals="proposals"
-        show-space
-        @end-reached="handleEndReached"
-      />
     </div>
-    <div class="invisible xl:visible fixed w-[240px] border-l bottom-0 top-[72px] right-0" />
+    <ProposalsList
+      title="Proposals"
+      limit="off"
+      :loading="!loaded"
+      :loading-more="loadingMore"
+      :proposals="proposals"
+      show-space
+      @end-reached="handleEndReached"
+    />
   </div>
 </template>
